[#45] generate_report를 별도 모듈로 분리
